# --list=build-info fixes