--- a/src/listconf.c
+++ b/src/listconf.c
@@ -188,14 +188,10 @@ static void listconf_list_build_info(void)
 	printf("$JOHN is %s\n", path_expand("$JOHN/"));
 	printf("Format interface version: %d\n", FMT_MAIN_VERSION);
 	printf("Max. number of reported tunable costs: %d\n", FMT_TUNABLE_COSTS);
-	puts("Rec file version: " RECOVERY_V);
-	puts("Charset file version: " CHARSET_V);
-	printf("CHARSET_MIN: %d (0x%02x)\n", CHARSET_MIN, CHARSET_MIN);
-	printf("CHARSET_MAX: %d (0x%02x)\n", CHARSET_MAX, CHARSET_MAX);
-	printf("CHARSET_LENGTH: %d\n", CHARSET_LENGTH);
+	puts("Rec file version: " RECOVERY_V ", charset file version: " CHARSET_V);
+	printf("CHARSET_MIN: %d (0x%02x), CHARSET_MAX: %d (0x%02x), CHARSET_LENGTH: %d\n", CHARSET_MIN, CHARSET_MIN, CHARSET_MAX, CHARSET_MAX, CHARSET_LENGTH);
 	printf("SALT_HASH_SIZE: %u\n", SALT_HASH_SIZE);
-	printf("SINGLE_IDX_MAX: %u\n", SINGLE_IDX_MAX);
-	printf("SINGLE_BUF_MAX: %u\n", SINGLE_BUF_MAX);
+	printf("SINGLE_IDX_MAX: %u, SINGLE_BUF_MAX: %u\n", SINGLE_IDX_MAX, SINGLE_BUF_MAX);
 	printf("Effective limit: ");
 	if (sizeof(SINGLE_KEYS_TYPE) < 4 || sizeof(SINGLE_KEYS_UTYPE) < 4) {
 		if (SINGLE_MAX_WORDS(125) < SINGLE_MAX_WORDS(16))
@@ -205,8 +201,7 @@ static void listconf_list_build_info(void)
 			printf("Max. KPC %d\n", SINGLE_MAX_WORDS(125));
 	} else
 		printf("Number of salts vs. SingleMaxBufferSize\n");
-	printf("Max. Markov mode level: %d\n", MAX_MKV_LVL);
-	printf("Max. Markov mode password length: %d\n", MAX_MKV_LEN);
+	printf("Markov mode max. level: %d, length %d\n", MAX_MKV_LVL, MAX_MKV_LEN);
 
 #if __ICC
 	printf("icc version: %d.%d.%d (gcc %d.%d.%d compatibility)\n",
@@ -249,11 +244,6 @@ static void listconf_list_build_info(void)
 	printf("OpenCL headers version: %s\n",get_opencl_header_version());
 #endif
 #if HAVE_LIBCRYPTO
-	printf("Crypto library: OpenSSL\n");
-#else
-	printf("Crypto library: None\n");
-#endif
-
 #if HAVE_LIBDL && defined(RTLD_DEFAULT)
 
 #if defined SSLEAY_VERSION && !defined OPENSSL_VERSION
@@ -262,8 +252,22 @@ static void listconf_list_build_info(void)
 #define SSLEAY_VERSION OPENSSL_VERSION
 #endif
 
-#ifdef OPENSSL_VERSION_NUMBER
-	printf("OpenSSL library version: %09lx", (unsigned long)OPENSSL_VERSION_NUMBER);
+#ifdef OPENSSL_VERSION_TEXT
+	printf("Crypto library: %s", OPENSSL_VERSION_TEXT);
+	if (dlsym(RTLD_DEFAULT, "OpenSSL_version")) {
+		const char* (*OpenSSL_version)(int) =
+			dlsym(RTLD_DEFAULT, "OpenSSL_version");
+
+		if (strcmp(OPENSSL_VERSION_TEXT, OpenSSL_version(OPENSSL_VERSION)))
+			printf("\t(loaded: %s)", OpenSSL_version(OPENSSL_VERSION));
+	} else if (dlsym(RTLD_DEFAULT, "SSLeay_version")) {
+		const char* (*SSLeay_version)(int) = dlsym(RTLD_DEFAULT, "SSLeay_version");
+		if (strcmp(OPENSSL_VERSION_TEXT, SSLeay_version(SSLEAY_VERSION)))
+			printf("\t(loaded: %s)", SSLeay_version(SSLEAY_VERSION));
+	}
+	printf("\n");
+#elif defined OPENSSL_VERSION_NUMBER
+	printf("Crypto library: OpenSSL version: %09lx", (unsigned long)OPENSSL_VERSION_NUMBER);
 	if (dlsym(RTLD_DEFAULT, "OpenSSL_version_num")) {
 		unsigned long (*OpenSSL_version_num)(void) =
 			dlsym(RTLD_DEFAULT, "OpenSSL_version_num");
@@ -278,22 +282,12 @@ static void listconf_list_build_info(void)
 	}
 	printf("\n");
 #endif
-#ifdef OPENSSL_VERSION_TEXT
-	printf("%s", OPENSSL_VERSION_TEXT);
-	if (dlsym(RTLD_DEFAULT, "OpenSSL_version")) {
-		const char* (*OpenSSL_version)(int) =
-			dlsym(RTLD_DEFAULT, "OpenSSL_version");
-
-		if (strcmp(OPENSSL_VERSION_TEXT, OpenSSL_version(OPENSSL_VERSION)))
-			printf("\t(loaded: %s)", OpenSSL_version(OPENSSL_VERSION));
-	} else if (dlsym(RTLD_DEFAULT, "SSLeay_version")) {
-		const char* (*SSLeay_version)(int) = dlsym(RTLD_DEFAULT, "SSLeay_version");
-		if (strcmp(OPENSSL_VERSION_TEXT, SSLeay_version(SSLEAY_VERSION)))
-			printf("\t(loaded: %s)", SSLeay_version(SSLEAY_VERSION));
-	}
-	printf("\n");
-#endif
+#else
+	printf("Crypto library: OpenSSL\n");
 #endif /* HAVE_LIBDL && defined(RTLD_DEFAULT) */
+#else
+	printf("Crypto library: None\n");
+#endif
 
 #ifdef __GNU_MP_VERSION
 	printf("GMP library version: %d.%d.%d",
@@ -322,15 +316,12 @@ static void listconf_list_build_info(void)
 #else
 	puts("File locking: NOT supported by this build - do not run concurrent sessions!");
 #endif
-	printf("fseek(): " STR_MACRO(jtr_fseek64) "\n");
-	printf("ftell(): " STR_MACRO(jtr_ftell64) "\n");
-	printf("fopen(): " STR_MACRO(jtr_fopen) "\n");
 #if HAVE_MEMMEM
 #define memmem_func	"System's"
 #else
 #define memmem_func	"JtR internal"
 #endif
-	printf("memmem(): " memmem_func "\n");
+	printf("fseek(): " STR_MACRO(jtr_fseek64) ", ftell(): " STR_MACRO(jtr_ftell64) ", fopen(): " STR_MACRO(jtr_fopen) ", memmem(): " memmem_func "\n");
 
 	clk_tck_init();
 #if defined(_SC_CLK_TCK) || !defined(CLK_TCK)