@claudioandre-br 2024-11-30
I'll take this opportunity: as there is no such thing as 'bleeding' non 'jumbo', in my opinion, only one of them is necessary.

Just say `bleeding`.