@solardiz 2024-12-01
> I'll take this opportunity: as there is no such thing as 'bleeding' non 'jumbo', in my opinion, only one of them is necessary.

Currently the word `jumbo` is simply part of our previous release number. The next release should be 2.0.0 without that in it, at which time the string reported here will also become the way you like.